Create helloWorld.java
This is the first basic program in java with just a simple class and main function. It just prints hello world on the console. To run this program, there should be jdk recent version installed in your machine. A text editor like notepad or an ide like eclipse can be used to run this program. In this scenario, I will show how to run the above code with a text editor i.e. Notepad in Windows. Follow the steps below:
1. Open a text editor and paste the above code.
2. Save the file as helloWorld.java (Note: file name should be same as the public class name given in the program)
3. Open windows command prompt(cmd) and type this code 'javac helloWorld.java'
4. This javac command will compile the given java file and produce a bytecode (helloWorld.class file) as output.
5. After compilation, type this code 'java helloWorld'  (Note: There is no need to append .java extension in this code)
6. This command prompts the jvm to run the bytecode of the helloWorld program, and prints Hello World! on the console as expected.
7. In this process if you see errors such as "javac is not recognized as an internal/external command....." then check if jdk is installed on your machine. If its available, then the path needs to be set. You can do this in the command prompt directly by typing "set path = C:\Program Files\Java\jdk1.8.0_221\bin" or you can edit the path variable in your system environment variables to the aforementioned path. Also you should add a new system variable JAVA_HOME and set its value to "C:\Program Files\Java\jdk1.8.0_221"